Refactor localization and settings structure; migrate to new feature directories, update dependencies, and enhance splash screen implementation
