+++ lib/src/features/home/presentation/homescreen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:league_of_logs/src/settings/settings_view.dart';
-import 'package:league_of_logs/src/utils/constants.dart';
-import 'package:league_of_logs/src/workout_generator/workout_generator_form.dart';
+import 'package:league_of_logs/src/features/settings/presentation/settings_view.dart';
+import 'package:league_of_logs/src/core/constants.dart';
+import 'package:league_of_logs/src/features/workout_generator/presentation/workout_generator_view.dart';
 
 class HomeScreen extends StatefulWidget {
   @override
