+++ lib/src/features/home/presentation/splashscreen.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:league_of_logs/src/home/homescreen.dart';
-import 'package:league_of_logs/src/utils/constants.dart';
+import 'package:league_of_logs/src/features/home/presentation/homescreen.dart';
+import 'package:league_of_logs/src/core/constants.dart';
 
 class SplashScreen extends StatefulWidget {
   @override
