+++ lib/src/features/workout_generator/data/player_stats.dart
