+++ lib/src/features/workout_generator/data/workout.dart
