+++ lib/src/features/workout_generator/presentation/workout_generator_view.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:league_of_logs/src/workout_generator/player_stats_DTO.dart';
+import 'package:league_of_logs/src/features/workout_generator/player_stats_DTO.dart';
 
-import '../settings/settings_view.dart';
-import 'package:league_of_logs/src/workout_generator/workout_generator_service.dart';
-import 'package:league_of_logs/src/utils/constants.dart';
+import '../../settings/presentation/settings_view.dart';
+import 'package:league_of_logs/src/features/workout_generator/domain/workout_generator_service.dart';
+import 'package:league_of_logs/src/core/constants.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 class PostMatchStatsForm extends StatefulWidget {
