+++ test/workout_generator_service_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:league_of_logs/src/utils/constants.dart';
-import 'package:league_of_logs/src/workout_generator/player_stats_DTO.dart';
-import 'package:league_of_logs/src/workout_generator/workout_generator_service.dart';
+import 'package:league_of_logs/src/core/constants.dart';
+import 'package:league_of_logs/src/features/workout_generator/player_stats_DTO.dart';
+import 'package:league_of_logs/src/features/workout_generator/domain/workout_generator_service.dart';
 
 void main() {
   group('WorkoutGeneratorService', () {
